WIP Settings. Refactored GridScreen
