+++ app/src/main/java/com/softyorch/famousquotes/ui/core/commonComponents/SpacerIconButton.kt
@@ -1,0 +1,13 @@
+package com.softyorch.famousquotes.ui.core.commonComponents
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun SpacerIconButton(modifier: Modifier = Modifier) {
+    Box(modifier = modifier.padding(4.dp).size(48.dp))
+}
