+++ app/src/main/java/com/softyorch/famousquotes/ui/core/navigation/Routes.kt
@@ -10,3 +10,6 @@ object Home
 
 @Serializable
 object Grid
+
+@Serializable
+object Settings
