+++ app/src/main/java/com/softyorch/famousquotes/ui/screens/grid/components/ButtonTopBar.kt
@@ -1,0 +1,46 @@
+package com.softyorch.famousquotes.ui.screens.grid.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.unit.dp
+import com.softyorch.famousquotes.core.FilterQuotes
+import com.softyorch.famousquotes.ui.core.commonComponents.IconButtonMenu
+import com.softyorch.famousquotes.ui.theme.BackgroundColor
+
+@Composable
+fun ButtonTopBar(
+    contentDescription: String,
+    filterQuotes: FilterQuotes,
+    icon: ImageVector,
+    isSelected: Boolean,
+    onClickListener: (FilterQuotes) -> Unit
+) {
+    Box(
+        modifier = Modifier
+            .padding(horizontal = 1.dp)
+            .width(80.dp)
+            .height(40.dp)
+            .shadow(elevation = 4.dp, shape = MaterialTheme.shapes.large)
+            .background(
+                color = BackgroundColor,
+                shape = MaterialTheme.shapes.large
+            ),
+        contentAlignment = Alignment.BottomCenter
+    ) {
+        IconButtonMenu(
+            cDescription = contentDescription,
+            icon = icon,
+            shadowOn = false,
+            isSelected = isSelected
+        ) { onClickListener(filterQuotes) }
+    }
+}
