+++ app/src/main/java/com/softyorch/famousquotes/ui/screens/grid/components/CardItem.kt
@@ -1,0 +1,101 @@
+package com.softyorch.famousquotes.ui.screens.grid.components
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImagePainter
+import coil.compose.rememberAsyncImagePainter
+import coil.request.ImageRequest
+import com.softyorch.famousquotes.R
+import com.softyorch.famousquotes.domain.model.FamousQuoteModel
+import com.softyorch.famousquotes.ui.components.IsDebugShowText
+import com.softyorch.famousquotes.ui.theme.WhiteSmoke
+
+@Composable
+fun CardItem(item: FamousQuoteModel?) {
+    val context = LocalContext.current
+
+    Card(
+        modifier = Modifier
+            .padding(start = 4.dp, end = 4.dp, bottom = 16.dp)
+            .fillMaxWidth()
+            .height(240.dp),
+        shape = MaterialTheme.shapes.large,
+        colors = CardDefaults.cardColors(
+            containerColor = WhiteSmoke
+        ),
+        elevation = CardDefaults.elevatedCardElevation(defaultElevation = 4.dp),
+    ) {
+        Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.TopCenter) {
+            if (item == null || item.imageUrl.isEmpty()) {
+                Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
+                    CircularProgressIndicator()
+                }
+            } else {
+                val painter = rememberAsyncImagePainter(
+                    model = ImageRequest.Builder(context)
+                        .data(item.imageUrl)
+                        .crossfade(true) // Set the target size to load the image at.
+                        .build(),
+                    contentScale = ContentScale.Fit
+                )
+
+                val painterState = painter.state
+
+                Column(
+                    modifier = Modifier.fillMaxSize(),
+                    horizontalAlignment = Alignment.CenterHorizontally,
+                    verticalArrangement = Arrangement.Center
+                ) {
+                    Card(
+                        modifier = Modifier
+                            .height(180.dp)
+                            .padding(4.dp),
+                        shape = MaterialTheme.shapes.large,
+                        elevation = CardDefaults.elevatedCardElevation(defaultElevation = 2.dp),
+                    ) {
+                        Image(
+                            painter = painter,
+                            contentDescription = stringResource(R.string.main_content_desc_image),
+                            modifier = Modifier
+                                .fillMaxWidth()
+                                .height(180.dp),
+                            contentScale = ContentScale.Crop
+                        )
+                    }
+
+                    Text(
+                        text = item.body,
+                        style = MaterialTheme.typography.bodyMedium,
+                        modifier = Modifier.padding(8.dp),
+                    )
+                }
+
+                IsDebugShowText(item.id)
+
+                if (painterState !is AsyncImagePainter.State.Success) {
+                    Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
+                        CircularProgressIndicator()
+                    }
+                }
+            }
+        }
+    }
+}
