+++ app/src/main/java/com/softyorch/famousquotes/ui/screens/settings/SettingsViewModel.kt
@@ -1,0 +1,10 @@
+package com.softyorch.famousquotes.ui.screens.settings
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class SettingsViewModel @Inject constructor(): ViewModel() {
+
+}
